transaction user verification

+++ api/controllers/transaction/delete.js
@@ -1,5 +1,3 @@
-const Category = require('../../models/Transaction');
-
 module.exports = {
 
 
@@ -23,12 +21,23 @@ module.exports = {
     notFound: {
       description: 'No transaction with the specified ID was found in the database.',
       responseType: 'notFound'
+    },
+    notPermitted: {
+      description: 'You are not permitted to see this record.',
+      responseType: 'notpermitted'
     }
   },
 
   fn: async function ({ transactionId }) {
+    var transaction = await Transaction.findOne({ id: transactionId });
+
+    if (!transaction) { throw 'notFound'; }
+    if (transaction.user !== this.req.session.userId) {
+      throw 'notPermitted';
+    }
+
     await Transaction.destroy({ id: transactionId });
-  
+
     return {
       response: 'Transaction was deleted'
     };
